Refactor policies page
