+++ studio/components/interfaces/Auth/Policies/index.tsx
@@ -1,7 +1,8 @@
+import Policies from './Policies'
 import PolicyEditorModal from './PolicyEditorModal'
 import PolicyTableRow from './PolicyTableRow'
 import PolicySelection from './PolicySelection'
 import PolicyTemplates from './PolicyTemplates'
 
 export * from './Policies.constants'
-export { PolicyEditorModal, PolicyTableRow, PolicySelection, PolicyTemplates }
+export { Policies, PolicyEditorModal, PolicyTableRow, PolicySelection, PolicyTemplates }
